added user authentication with Sanctum, created AuthController, Task model eloquent relations, added authentication and autorization mechanisms to protect certain endpoints in TaskController  and protected certain api routes

+++ app/Http/Controllers/TaskController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Task;
+use App\Models\User;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -27,7 +28,9 @@ class TaskController extends Controller
     public function store(Request $request)
     {
         $this->validateTask($request);
-        return Task::create($request->all());
+        $taskData = $request->all();
+        $taskData['creator_id'] = auth()->id();
+        return Task::create($taskData);
     }
 
     /**
@@ -48,27 +51,58 @@ class TaskController extends Controller
      */
     public function update(Request $request, string $id)
     {
+
+
         $this->validateTask($request);
         $task = Task::find($id);
+
         if (!$task) {
             return response()->json(['message' => 'Task not found'], 404);
         }
-        $task->update($request->all());
+
+        if($task->creator->id !== auth()->id() && $task->assignedTo->id !== auth()->id()) {
+            return response()->json(['message' => 'Only creator and assigned user can change task!'], 403);
+        }
+
+        $taskData = $request->except('creator_id');
+
+        if($task->assignedTo->id === auth()->id()){
+            unset($taskData['assigned_user_id']);
+        }
+
+
+        $task->update($taskData);
         return response()->json(['message' => 'Task updated successfully']);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $id, User $user)
     {
+
         try {
             $task = Task::findOrFail($id);
-            $task->delete();
-            return response()->json(['message' => 'Task deleted successfully']);
+            if($task->creator->id === auth()->id() || auth()->user()->role === 'Admin'){
+                $task->delete();
+                return response()->json(['message' => 'Task deleted successfully']);
+            } else return response()->json(['message' => 'Only task creator and admin can delete task!'], 403);
+
         } catch (ModelNotFoundException $e) {
             return response()->json(['message' => 'Task not found'], 404);
         }
+    }
+
+    /**
+     * Search for task by title
+     */
+    public function search($title)
+    {
+        $tasks = Task::where('title', 'like', '%'.$title.'%')->get();
+        if ($tasks->isEmpty()) {
+            return response()->json(['message' => 'No task found'], 404);
+        }
+        return $tasks;
     }
 
     /**
@@ -81,8 +115,8 @@ class TaskController extends Controller
             'status' => ['required', Rule::in(['New', 'In Progress', 'Completed'])],
             'priority' => ['required', Rule::in(['Low', 'Medium', 'High'])],
             'due_date' => 'required|date_format:Y-m-d H:i:s',
-            'creator_id' => 'required',
-            'assigned_user_id' => 'required'
+            'creator_id' => ['nullable', Rule::exists('users', 'id')],
+            'assigned_user_id' => ['required', Rule::exists('users', 'id')],
         ]);
     }
 }
